LPS-95267 No need for two utils
